working empty hidra implementation
Strings representing the names of the methods are sent and checked. The ACS handles only one exchange at a time, initiated by the Subject. The resources handle 1 Hidra exchange at the same time, which can be repeated, but internally this is not remembered. Steps 8 and 9, the accounting steps, are not implemented.